+++ src/CLR/WireProtocol/WireProtocol_MonitorCommands.c
@@ -5,50 +5,102 @@
 
 #include <nanoWeak.h>
 #include "WireProtocol_MonitorCommands.h"
+#include <nanoPAL_BlockStorage.h>
 #include <target_board.h>
 
-////////////////////////////////////////////////////
+#if !defined(_WIN32)
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_Ping(WP_Message *message)
-{
-    (void)(message);
+//////////////////////////////////////////////////
 
-    // default to false
-    return false;
-}
-
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_OemInfo(WP_Message *message)
+int Monitor_Ping(WP_Message *message)
 {
     if ((message->m_header.m_flags & WP_Flags_c_Reply) == 0)
     {
-        Monitor_OemInfo_Reply cmdReply;
+        Monitor_Ping_Reply cmdReply;
+        cmdReply.Source = Monitor_Ping_c_Ping_Source_NanoBooter;
+        cmdReply.Flags = 0;
 
-        bool fOK = NanoBooter_GetReleaseInfo(&cmdReply.m_releaseInfo) == true;
+        // fill in the flags
+#if defined(WP_IMPLEMENTS_CRC32)
+        cmdReply.Flags |= Monitor_Ping_c_Ping_WPFlag_SupportsCRC32;
+#endif
 
-        WP_ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply));
+// Wire Protocol packet size
+#if (WP_PACKET_SIZE == 512)
+        cmdReply.Flags |= Monitor_Ping_c_PacketSize_0512;
+#elif (WP_PACKET_SIZE == 256)
+        cmdReply.Flags |= Monitor_Ping_c_PacketSize_0256;
+#elif (WP_PACKET_SIZE == 128)
+        cmdReply.Flags |= Monitor_Ping_c_PacketSize_0128;
+#elif (WP_PACKET_SIZE == 1024)
+        cmdReply.Flags |= Monitor_Ping_c_PacketSize_1024;
+#endif
+
+        if (Target_HasNanoBooter())
+        {
+            cmdReply.Flags |= Monitor_Ping_c_HasNanoBooter;
+        }
+
+        if (Target_HasProprietaryBooter())
+        {
+            cmdReply.Flags |= Monitor_Ping_c_HasProprietaryBooter;
+        }
+
+        if (Target_IFUCapable())
+        {
+            cmdReply.Flags |= Monitor_Ping_c_IFUCapable;
+        }
+
+        if (Target_ConfigUpdateRequiresErase())
+        {
+            cmdReply.Flags |= Monitor_Ping_c_ConfigBlockRequiresErase;
+        }
+
+        WP_ReplyToCommand(message, true, false, &cmdReply, sizeof(cmdReply));
     }
 
     return true;
 }
 
 // provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_ReadMemory(WP_Message *message)
+__nfweak int Monitor_OemInfo(WP_Message *message)
 {
-    (void)(message);
+    Monitor_OemInfo_Reply cmdReply;
 
-    // default to false
-    return false;
+    bool fOK = NanoBooter_GetReleaseInfo(&cmdReply.m_releaseInfo) == true;
+
+    WP_ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply));
+
+    return true;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_WriteMemory(WP_Message *message)
+int Monitor_ReadMemory(WP_Message *message)
 {
-    (void)(message);
+    CLR_DBG_Commands_Monitor_ReadMemory *cmd = (CLR_DBG_Commands_Monitor_ReadMemory *)message->m_payload;
 
-    // default to false
-    return false;
+    unsigned char buf[1024];
+    unsigned int len = cmd->length;
+    if (len > sizeof(buf))
+        len = sizeof(buf);
+    uint32_t errorCode;
+
+    AccessMemory(cmd->address, len, buf, AccessMemory_Read, &errorCode);
+
+    WP_ReplyToCommand(message, true, false, buf, len);
+
+    return true;
+}
+
+int Monitor_WriteMemory(WP_Message *message)
+{
+    CLR_DBG_Commands_Monitor_WriteMemory *cmd = (CLR_DBG_Commands_Monitor_WriteMemory *)message->m_payload;
+    CLR_DBG_Commands_Monitor_WriteMemory_Reply cmdReply;
+
+    AccessMemory(cmd->address, cmd->length, cmd->data, AccessMemory_Write, &cmdReply);
+
+    WP_ReplyToCommand(message, true, false, &cmdReply, sizeof(cmdReply));
+
+    return true;
 }
 
 // provided as weak to be replaced by actual implementation by application
@@ -60,71 +112,245 @@ __nfweak int Monitor_Reboot(WP_Message *message)
     return false;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_EraseMemory(WP_Message *message)
+int Monitor_EraseMemory(WP_Message *message)
 {
-    (void)(message);
+    CLR_DBG_Commands_Monitor_EraseMemory *cmd = (CLR_DBG_Commands_Monitor_EraseMemory *)message->m_payload;
+    CLR_DBG_Commands_Monitor_EraseMemory_Reply cmdReply;
 
-    // default to false
-    return false;
+    // TODO: not sure if we really need this
+    // nanoBooter_OnStateChange( State_MemoryErase, (void*)cmd->m_address );
+
+    AccessMemory(cmd->address, cmd->length, NULL, AccessMemory_Erase, &cmdReply);
+
+    WP_ReplyToCommand(message, true, false, &cmdReply, sizeof(cmdReply));
+
+    return true;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_QueryConfiguration(WP_Message *message)
+int Monitor_QueryConfiguration(WP_Message *message)
 {
-    (void)(message);
+    bool success = false;
 
-    // default to false
-    return false;
+    // include handling of configuration block only if feature is available
+#if (HAS_CONFIG_BLOCK == TRUE)
+
+    Monitor_QueryConfiguration_Command *cmd = (Monitor_QueryConfiguration_Command *)message->m_payload;
+    int size = 0;
+
+    HAL_Configuration_NetworkInterface configNetworkInterface;
+    HAL_Configuration_Wireless80211 configWireless80211NetworkInterface;
+
+    switch ((DeviceConfigurationOption)cmd->Configuration)
+    {
+        case DeviceConfigurationOption_Network:
+
+            if (ConfigurationManager_GetConfigurationBlock(
+                    (void *)&configNetworkInterface,
+                    (DeviceConfigurationOption)cmd->Configuration,
+                    cmd->BlockIndex) == true)
+            {
+                size = sizeof(HAL_Configuration_NetworkInterface);
+                success = true;
+
+                WP_ReplyToCommand(message, success, false, (uint8_t *)&configNetworkInterface, size);
+            }
+            break;
+
+        case DeviceConfigurationOption_Wireless80211Network:
+
+            if (ConfigurationManager_GetConfigurationBlock(
+                    (void *)&configWireless80211NetworkInterface,
+                    (DeviceConfigurationOption)cmd->Configuration,
+                    cmd->BlockIndex) == true)
+            {
+                size = sizeof(HAL_Configuration_Wireless80211);
+                success = true;
+
+                WP_ReplyToCommand(message, success, false, (uint8_t *)&configWireless80211NetworkInterface, size);
+            }
+            break;
+
+        case DeviceConfigurationOption_WirelessNetworkAP:
+            // TODO missing implementation for now
+            break;
+
+        default:
+            break;
+    }
+
+    if (!success)
+    {
+        WP_ReplyToCommand(message, success, false, NULL, size);
+    }
+
+#else
+
+    (void)message;
+
+#endif // (HAS_CONFIG_BLOCK == TRUE)
+
+    return success;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_UpdateConfiguration(WP_Message *message)
+int Monitor_UpdateConfiguration(WP_Message *message)
 {
-    (void)(message);
+    bool success = false;
 
-    // default to false
-    return false;
+    // include handling of configuration block only if feature is available
+#if (HAS_CONFIG_BLOCK == TRUE)
+
+    Monitor_UpdateConfiguration_Command *cmd = (Monitor_UpdateConfiguration_Command *)message->m_payload;
+    Monitor_UpdateConfiguration_Reply cmdReply;
+
+    switch ((DeviceConfigurationOption)cmd->Configuration)
+    {
+        case DeviceConfigurationOption_Network:
+        case DeviceConfigurationOption_Wireless80211Network:
+        case DeviceConfigurationOption_X509CaRootBundle:
+        case DeviceConfigurationOption_All:
+            if (ConfigurationManager_StoreConfigurationBlock(
+                    cmd->Data,
+                    (DeviceConfigurationOption)cmd->Configuration,
+                    cmd->BlockIndex,
+                    cmd->Length,
+                    cmd->Offset,
+                    cmd->Done) == true)
+            {
+                cmdReply.ErrorCode = 0;
+                success = true;
+            }
+            else
+            {
+                cmdReply.ErrorCode = 100;
+            }
+            break;
+
+        default:
+            cmdReply.ErrorCode = 10;
+    }
+
+    WP_ReplyToCommand(message, success, false, &cmdReply, sizeof(cmdReply));
+
+#else
+
+    (void)message;
+
+#endif // (HAS_CONFIG_BLOCK == TRUE)
+
+    return success;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_CheckMemory(WP_Message *message)
+int Monitor_CheckMemory(WP_Message *message)
 {
-    (void)(message);
+    bool ret = false;
 
-    // default to false
-    return false;
+    CLR_DBG_Commands_Monitor_CheckMemory *cmd = (CLR_DBG_Commands_Monitor_CheckMemory *)message->m_payload;
+    CLR_DBG_Commands_Monitor_CheckMemory_Reply cmdReply;
+    uint32_t errorCode;
+
+    ret = AccessMemory(cmd->address, cmd->length, (uint8_t *)&cmdReply, AccessMemory_Check, &errorCode);
+
+    WP_ReplyToCommand(message, ret, false, &cmdReply, sizeof(cmdReply));
+
+    return ret;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_MemoryMap(WP_Message *message)
+int Monitor_MemoryMap(WP_Message *message)
 {
-    (void)(message);
+    MemoryMap_Range map[2];
 
-    // default to false
-    return false;
+    map[0].m_address = HalSystemConfig.RAM1.Base;
+    map[0].m_length = HalSystemConfig.RAM1.Size;
+    map[0].m_flags = Monitor_MemoryMap_c_RAM;
+
+    map[1].m_address = HalSystemConfig.FLASH1.Base;
+    map[1].m_length = HalSystemConfig.FLASH1.Size;
+    map[1].m_flags = Monitor_MemoryMap_c_FLASH;
+
+    WP_ReplyToCommand(message, true, false, map, sizeof(map));
+
+    return true;
 }
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_FlashSectorMap(WP_Message *message)
-{
-    (void)(message);
-
-    // default to false
-    return false;
-}
-
-// provided as weak to be replaced by actual implementation by application
-__nfweak int Monitor_TargetInfo(WP_Message *message)
+int Monitor_FlashSectorMap(WP_Message *message)
 {
     if ((message->m_header.m_flags & WP_Flags_c_Reply) == 0)
     {
-        Monitor_TargetInfo_Reply cmdReply;
+        struct Flash_BlockRegionInfo
+        {
+            unsigned int StartAddress;
+            unsigned int NumBlocks;
+            unsigned int BytesPerBlock;
+            unsigned int Usage;
 
-        bool fOK = nanoBooter_GetTargetInfo(&cmdReply.m_TargetInfo) == true;
+        } *pData = NULL;
 
-        WP_ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply));
+        unsigned int rangeCount = 0;
+        unsigned int rangeIndex = 0;
+
+        for (int cnt = 0; cnt < 2; cnt++)
+        {
+            BlockStorageDevice *device = BlockStorageList_GetFirstDevice();
+
+            if (device == NULL)
+            {
+                WP_ReplyToCommand(message, true, false, NULL, 0);
+                return false;
+            }
+
+            if (cnt == 1)
+            {
+                pData =
+                    (struct Flash_BlockRegionInfo *)platform_malloc(rangeCount * sizeof(struct Flash_BlockRegionInfo));
+
+                if (pData == NULL)
+                {
+                    WP_ReplyToCommand(message, true, false, NULL, 0);
+                    return false;
+                }
+            }
+
+            DeviceBlockInfo *deviceInfo = BlockStorageDevice_GetDeviceInfo(device);
+
+            for (unsigned int i = 0; i < deviceInfo->NumRegions; i++)
+            {
+                const BlockRegionInfo *pRegion = &deviceInfo->Regions[i];
+
+                for (unsigned int j = 0; j < pRegion->NumBlockRanges; j++)
+                {
+
+                    if (cnt == 0)
+                    {
+                        rangeCount++;
+                    }
+                    else
+                    {
+                        pData[rangeIndex].StartAddress =
+                            BlockRegionInfo_BlockAddress(pRegion, pRegion->BlockRanges[j].StartBlock);
+                        pData[rangeIndex].NumBlocks = BlockRange_GetBlockCount(pRegion->BlockRanges[j]);
+                        pData[rangeIndex].BytesPerBlock = pRegion->BytesPerBlock;
+                        pData[rangeIndex].Usage = pRegion->BlockRanges[j].RangeType & BlockRange_USAGE_MASK;
+                        rangeIndex++;
+                    }
+                }
+            }
+        }
+
+        WP_ReplyToCommand(message, true, false, (void *)pData, rangeCount * sizeof(struct Flash_BlockRegionInfo));
+
+        platform_free(pData);
     }
+
+    return true;
+}
+
+int Monitor_TargetInfo(WP_Message *message)
+{
+    Monitor_TargetInfo_Reply cmdReply;
+
+    bool fOK = nanoBooter_GetTargetInfo(&cmdReply.m_TargetInfo) == true;
+
+    WP_ReplyToCommand(message, fOK, false, &cmdReply, sizeof(cmdReply));
 
     return true;
 }
@@ -132,8 +358,7 @@ __nfweak int Monitor_TargetInfo(WP_Message *message)
 //////////////////////////////////////////////////////////////////////
 // helper functions
 
-// provided as weak to be replaced by actual implementation by application
-__nfweak int nanoBooter_GetTargetInfo(TargetInfo *targetInfo)
+int nanoBooter_GetTargetInfo(TargetInfo *targetInfo)
 {
 #if (TARGET_HAS_NANOBOOTER == TRUE)
 
@@ -171,7 +396,7 @@ __nfweak int nanoBooter_GetTargetInfo(TargetInfo *targetInfo)
     return true;
 }
 
-__nfweak int NanoBooter_GetReleaseInfo(ReleaseInfo *releaseInfo)
+int NanoBooter_GetReleaseInfo(ReleaseInfo *releaseInfo)
 {
     releaseInfo->version.usMajor = VERSION_MAJOR;
     releaseInfo->version.usMinor = VERSION_MINOR;
@@ -185,3 +410,5 @@ __nfweak int NanoBooter_GetReleaseInfo(ReleaseInfo *releaseInfo)
 
     return true;
 }
+
+#endif // _WIN32
